--- a/control/kern/tproxy.c
+++ b/control/kern/tproxy.c
@@ -1620,10 +1620,23 @@ int local_tcp_sockops(struct bpf_sock_ops *skops)
 	tuple.l4proto = IPPROTO_TCP;
 	tuple.sport = bpf_htonl(skops->local_port) >> 16;
 	tuple.dport = skops->remote_port >> 16;
-	tuple.sip.u6_addr32[2] = bpf_htonl(0x0000ffff);
-	tuple.sip.u6_addr32[3] = skops->local_ip4;
-	tuple.dip.u6_addr32[2] = bpf_htonl(0x0000ffff);
-	tuple.dip.u6_addr32[3] = skops->remote_ip4;
+	if (skops->family == AF_INET) {
+		tuple.sip.u6_addr32[2] = bpf_htonl(0x0000ffff);
+		tuple.sip.u6_addr32[3] = skops->local_ip4;
+		tuple.dip.u6_addr32[2] = bpf_htonl(0x0000ffff);
+		tuple.dip.u6_addr32[3] = skops->remote_ip4;
+	} else if (skops->family == AF_INET6) {
+		tuple.sip.u6_addr32[3] = skops->local_ip6[3];
+		tuple.sip.u6_addr32[2] = skops->local_ip6[2];
+		tuple.sip.u6_addr32[1] = skops->local_ip6[1];
+		tuple.sip.u6_addr32[0] = skops->local_ip6[0];
+		tuple.dip.u6_addr32[3] = skops->remote_ip6[3];
+		tuple.dip.u6_addr32[2] = skops->remote_ip6[2];
+		tuple.dip.u6_addr32[1] = skops->remote_ip6[1];
+		tuple.dip.u6_addr32[0] = skops->remote_ip6[0];
+	} else {
+		return 0;
+	}
 
 	switch (skops->op) {
 	case BPF_SOCK_OPS_PASSIVE_ESTABLISHED_CB: // dae sockets
@@ -1671,10 +1684,23 @@ int sk_msg_fast_redirect(struct sk_msg_md *msg)
 	rev_tuple.l4proto = IPPROTO_TCP;
 	rev_tuple.sport = msg->remote_port >> 16;
 	rev_tuple.dport = bpf_htonl(msg->local_port) >> 16;
-	rev_tuple.sip.u6_addr32[2] = bpf_htonl(0x0000ffff);
-	rev_tuple.sip.u6_addr32[3] = msg->remote_ip4;
-	rev_tuple.dip.u6_addr32[2] = bpf_htonl(0x0000ffff);
-	rev_tuple.dip.u6_addr32[3] = msg->local_ip4;
+	if (msg->family == AF_INET) {
+		rev_tuple.sip.u6_addr32[2] = bpf_htonl(0x0000ffff);
+		rev_tuple.sip.u6_addr32[3] = msg->remote_ip4;
+		rev_tuple.dip.u6_addr32[2] = bpf_htonl(0x0000ffff);
+		rev_tuple.dip.u6_addr32[3] = msg->local_ip4;
+	} else if (msg->family == AF_INET6) {
+		rev_tuple.sip.u6_addr32[3] = msg->remote_ip6[3];
+		rev_tuple.sip.u6_addr32[2] = msg->remote_ip6[2];
+		rev_tuple.sip.u6_addr32[1] = msg->remote_ip6[1];
+		rev_tuple.sip.u6_addr32[0] = msg->remote_ip6[0];
+		rev_tuple.dip.u6_addr32[3] = msg->local_ip6[3];
+		rev_tuple.dip.u6_addr32[2] = msg->local_ip6[2];
+		rev_tuple.dip.u6_addr32[1] = msg->local_ip6[1];
+		rev_tuple.dip.u6_addr32[0] = msg->local_ip6[0];
+	} else {
+		return SK_PASS;
+	}
 
 	if (bpf_msg_redirect_hash(msg, &fast_sock, &rev_tuple, BPF_F_INGRESS) == SK_PASS)
 		bpf_printk("tcp fast redirect: %pI4:%lu -> %pI4:%lu",