@mzz2017 2024-03-27
这个优化非常令人兴奋，这或许已经是当前 linux 系统下的最优性能方案。通过 socket 重定向直接将路径缩至最短，非常极致的优化！

针对这次优化，是否需要更高版本的内核？如果是，我们或许需要增加一些判断和提示（像之前的代码那样），以及更新一些文档。