@jschwinger233 2024-03-27
> 针对这次优化，是否需要更高版本的内核？如果是，我们或许需要增加一些判断和提示（像之前的代码那样），以及更新一些文档。

CI 测过了 5.10 貌似是好的。 dae 目前要求 >=5.8，我自己编译一个 5.8 试试