@jschwinger233 2024-03-27
> 针对这次优化，是否需要更高版本的内核？如果是，我们或许需要增加一些判断和提示（像之前的代码那样），以及更新一些文档。

编译了 5.8 （妈的这版本 EOL 了我手动改了 objtool/elf.c 才编过，还把我磁盘占满了），不能运行，报错 `in-kernel BTF is malformed`，但我觉得单纯是因为 5.8 又老又 EOL 在编译时 binutils 没有正确生成 BTF，不代表真的无法在 5.8 运行。

但是考虑到以后我可能很难测试 5.8，如果可以稍微提高内核要求到 5.10 就更好了，5.10 是一个 LTS 版本，要 31 Dec 2026 才停止支持 ( https://endoflife.date/linux ) ，目前的 CI Kernel-test 也有测它。