@amtoaer 2024-03-28
使用该版本 dae 遇到一个问题。抽象出来应该是这种情况：
在 dae 宿主机运行两个 docker 容器 A、B 提供 web 服务，均使用 network_mode: bridge 运行。其中 A 的端口映射为 a:a，B 的端口映射为 b:b。
docker 的默认 bridge 如下：
```
docker0: flags=4163<UP,BROADCAST,RUNNING,MULTICAST>  mtu 1500
        inet 172.17.0.1  netmask 255.255.0.0  broadcast 172.17.255.255
```
此时，在 A 容器中访问 http://172.17.0.1:b/ 理应能够访问到 B 容器的 web 服务，但使用该 PR 的 build，这个请求会无响应。

---
daily main，无论是否开启 dae 均对该类请求无影响：
![image](https://github.com/daeuniverse/dae/assets/32017007/a3ee0116-f0e6-4210-ba17-70a3f6cf1450)
该 PR，开启 dae 后请求无响应:
![image](https://github.com/daeuniverse/dae/assets/32017007/a43f12e8-62f1-4f10-9145-b658ff3efd0c)