@amtoaer 2024-03-28
@jschwinger233 是的，我的配置是：
```
    lan_interface: docker0,br0
    wan_interface: br0
```